LibGL: Change GLsizei from unsigned to signed integral type
There is some really wild stuff going on in the OpenGL spec for this..
The Khronos website states that GLsizei is a 32-bit non-negative value
used for sizes, however, some functions such as `glGenTextures` state
that the input `n` could be negative, which implies signage. Most other
implementations of `gl.h` seem to `typedef` this to `int` so we should
too.

+++ Userland/Libraries/LibGL/GL/gl.h
@@ -163,7 +163,7 @@ typedef unsigned int GLuint;
 typedef int GLfixed;
 typedef long long GLint64;
 typedef unsigned long long GLuint64;
-typedef unsigned long GLsizei;
+typedef int GLsizei;
 typedef void GLvoid;
 typedef float GLfloat;
 typedef float GLclampf;
